Extract gutters into helix-view
