+++ helix-view/src/view.rs
@@ -1,6 +1,10 @@
 use std::borrow::Cow;
 
-use crate::{graphics::Rect, Document, DocumentId, ViewId};
+use crate::{
+    graphics::Rect,
+    gutter::{self, Gutter},
+    Document, DocumentId, ViewId,
+};
 use helix_core::{
     graphemes::{grapheme_width, RopeGraphemes},
     line_ending::line_end_char_index,
@@ -60,6 +64,8 @@ impl JumpList {
     }
 }
 
+const GUTTERS: &[(Gutter, usize)] = &[(gutter::diagnostic, 1), (gutter::line_number, 5)];
+
 #[derive(Debug)]
 pub struct View {
     pub id: ViewId,
@@ -83,10 +89,19 @@ impl View {
         }
     }
 
+    pub fn gutters(&self) -> &[(Gutter, usize)] {
+        GUTTERS
+    }
+
     pub fn inner_area(&self) -> Rect {
-        // TODO: not ideal
-        const OFFSET: u16 = 7; // 1 diagnostic + 5 linenr + 1 gutter
-        self.area.clip_left(OFFSET).clip_bottom(1) // -1 for statusline
+        // TODO: cache this
+        let offset = self
+            .gutters()
+            .iter()
+            .map(|(_, width)| *width as u16)
+            .sum::<u16>()
+            + 1; // +1 for some space between gutters and line
+        self.area.clip_left(offset).clip_bottom(1) // -1 for statusline
     }
 
     //
@@ -276,6 +291,7 @@ mod tests {
     use super::*;
     use helix_core::Rope;
     const OFFSET: u16 = 7; // 1 diagnostic + 5 linenr + 1 gutter
+                           // const OFFSET: u16 = GUTTERS.iter().map(|(_, width)| *width as u16).sum();
 
     #[test]
     fn test_text_pos_at_screen_coords() {
